fix scrolling bug
despite my previous commit message, the bug with a scrolled page persisted.
now it should be gone via window.scrollTo(0,0)

this commit also refines scrolling back to index. when a navigation from index to an article occurs the offsetY value is now stored and can be retrieved later on when navigating back to index